Move april fools related stuff to meta.py
