+++ addons/utility.py
@@ -13,7 +13,6 @@ import qrcode
 import io
 import hashlib
 import validators
-import asyncio
 from addons.helper import restricted_to_bot
 
 
@@ -627,48 +626,6 @@ class Utility(commands.Cog):
             embed.add_field(name="Permitted Roles", value=f"`{'`, `'.join(r.name for r in emote.roles if not 'FlagBot' in r.name)}`", inline=False)
         embed.add_field(name="Added On", value=f"{emote.created_at.strftime('%m-%d-%Y %H:%M:%S')} UTC", inline=False)
         await ctx.send(embed=embed)
-
-    @commands.command()
-    @commands.has_any_role("Admin", "Bot Dev")
-    async def surprise(self, ctx):
-        """Happy April 1st! Pass in image attachment for server icon change"""
-        if not os.path.exists('saves/layout.json'):
-            data = {}
-            for c in ctx.guild.channels:
-                data[c.id] = c.name
-            with open('saves/layout.json', 'w') as f:
-                json.dump(data, f, indent=4)
-        with open('saves/new_layout.json', 'r') as f:
-            new = json.load(f)
-        for ch in new.keys():
-            channel = ctx.guild.get_channel(int(ch))
-            if channel.name == new[ch]:
-                continue
-            await channel.edit(name=new[ch])
-            await asyncio.sleep(3)
-        if bool(ctx.message.attachments):
-            img_bytes = await ctx.message.attachments[0].read()
-        await ctx.guild.edit(icon=img_bytes)
-        await ctx.send("Updated the channel list.")
-
-    @commands.command()
-    @commands.has_any_role("Admin", "Bot Dev")
-    async def revert(self, ctx):
-        """April 1st is over"""
-        with open('saves/layout.json', 'r') as f:
-            old = json.load(f)
-        for ch in old.keys():
-            channel = ctx.guild.get_channel(int(ch))
-            if channel.name == old[ch]:
-                continue
-            await channel.edit(name=old[ch])
-            await asyncio.sleep(3)
-        async with self.bot.session.get(url='https://avatars.githubusercontent.com/u/42673825?s=200&v=4') as r:
-            r_data = await r.read()
-            await ctx.guild.edit(icon=r_data)
-        await ctx.send("Reverted back the channel list.")
-        
-        
             
 
 def setup(bot):
